HSLdevcom/bultti#220 Refactor and clean up Table component.
